pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "gtts>=2.5.4",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[all]>=22.5",
    "sqlalchemy>=2.0.43",
]
